Add a basic session to the mix
This is incomplete but its a beginning. A session allows you to make dispatch request and receive a response using a shared socket. This is the whole point of mux.

Right now only simple dispatches are implemented. If you don't get back a Rdispatch frame, everything is broken! Hardening and implementing things like ping are the next goal.